implementa tratamento de excecões e regras de negócio

+++ src/main/java/com/github/carlosouza_dev/libraryapi/service/AutorService.java
@@ -1,7 +1,11 @@
 package com.github.carlosouza_dev.libraryapi.service;
 
+import com.github.carlosouza_dev.libraryapi.exception.OperacaoNaoPermitidaException;
 import com.github.carlosouza_dev.libraryapi.model.Autor;
 import com.github.carlosouza_dev.libraryapi.repository.AutorRepository;
+import com.github.carlosouza_dev.libraryapi.repository.LivroRepository;
+import com.github.carlosouza_dev.libraryapi.validator.AutorValidator;
+import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -9,21 +13,22 @@ import java.util.Optional;
 import java.util.UUID;
 
 @Service
+@RequiredArgsConstructor
 public class AutorService {
 
     private final AutorRepository repository;
-
-    public AutorService(AutorRepository repository){
-        this.repository = repository;
-    }
+    private final AutorValidator validator;
+    private final LivroRepository livroRepository;
 
     public Autor salvar(Autor autor){
+        validator.validar(autor);
         return repository.save(autor);
     }
 
     public void atualizar(Autor autor){
         if (autor.getId() == null) throw new IllegalArgumentException("Autor precisa ter id para atualizar");
 
+        validator.validar(autor);
         repository.save(autor);
     }
 
@@ -32,7 +37,13 @@ public class AutorService {
     }
 
     public void deletar(Autor autor){
+        if (possuiLivros(autor))
+            throw new OperacaoNaoPermitidaException("Não é possível excluir Autor(a) que possua livros cadastrados");
         repository.delete(autor);
+    }
+
+    private boolean possuiLivros(Autor autor) {
+        return livroRepository.existsByAutor(autor);
     }
 
     public List<Autor> filtrar(String nome, String nacionalidade){
